requirements.txt
numpy
ollama
colorama
streamlit
fastapi
uvicorn
pydantic
sentence-transformers
networkx
pandas
openpyxl
matplotlib
seaborn
PyJWT
bcrypt
cryptography
slowapi
